Feature/monitor service

+++ cpu-monitor-api/src/main/java/com/ruan/cpu/dto/ProcessInfoDTO.java
@@ -8,9 +8,9 @@ import lombok.*;
 @AllArgsConstructor
 
 public class ProcessInfoDTO {
-    private Long pid;
+    private long pid;
     private String name;
     private double cpuLoadPercent;
-    private byte memoryBytes;
+    private Long memoryBytes;
 
 }
